M4: multi-molecule/multi-rep scene with selections + icon panel UI + CLAUDE.md
scene.rs adds Scene/Molecule/Representation; each rep has a molar selection
string (compile_selection -> system.select) compiled to atom indices, with
geometry built per subset. Renderer is per-rep (RepGpu) via render_scene over the
whole scene. scene_dirty render-skip: the 3D pass runs only on camera/geometry/
visibility/size change (no continuous repaint -> idle = 0 GPU).

UI: Scene section with icon-only perspective/orthographic toggle buttons;
molecule and rep rows show the name with a right-justified icon group (eye/
eye-slash visibility, duplicate for reps, trash delete); 'Add representation'
button before the rep list; rep label is <sel>/<style>. Icons via egui-phosphor.

CPU-side indices are usize (no u32 round-trips). Added CLAUDE.md onboarding doc.

+++ crates/vmd_rs_core/src/data/bonds.rs
@@ -14,7 +14,7 @@ const MIN_DIST: f32 = 0.04;
 /// Guess bonds for all atoms. `sel` is the bound all-selection (the position
 /// source for the grid search); `positions`/`vdw` are the extracted per-atom
 /// arrays (nm) used to score candidate pairs.
-pub fn guess(sel: &impl PosProvider, positions: &[[f32; 3]], vdw: &[f32]) -> Vec<[u32; 2]> {
+pub fn guess(sel: &impl PosProvider, positions: &[[f32; 3]], vdw: &[f32]) -> Vec<[usize; 2]> {
     let n = positions.len();
     if n < 2 {
         return Vec::new();
@@ -24,7 +24,7 @@ pub fn guess(sel: &impl PosProvider, positions: &[[f32; 3]], vdw: &[f32]) -> Vec
         distance_search_single::<(usize, usize), Vec<_>>(SEARCH_CUTOFF, sel, 0..n);
 
     let min2 = MIN_DIST * MIN_DIST;
-    let mut bonds: Vec<[u32; 2]> = Vec::new();
+    let mut bonds: Vec<[usize; 2]> = Vec::new();
     for (i, j) in candidates {
         if i == j {
             continue;
@@ -39,7 +39,7 @@ pub fn guess(sel: &impl PosProvider, positions: &[[f32; 3]], vdw: &[f32]) -> Vec
         };
         let thresh = BOND_FACTOR * (vdw[a] + vdw[b]);
         if d2 > min2 && d2 < thresh * thresh {
-            bonds.push([a as u32, b as u32]);
+            bonds.push([a, b]);
         }
     }
